Don't try as hard to recover from crypto desync
Somewhat frequently the communication between PHEV and phev2mqtt will
desync in a weird way where the xor code gets malformed. Often this
occurs when the xor code becomes `00`. I would imagine that in this case
the xor code would have to be shifted or the next code be used or
something.

When this happens, it can be particularly difficult for phev2mqtt to
recover, as the code becomes always something else than anticipated.
Retrying 50 times before a reconnect is hardly of any benefit.

It is much easier to try less hard and "just" reconnect. If we could
have super reliable connectivity, that would be great, but in practice
even the app seems to have a hard time, so I don't have high hopes.
Instead we'll make the status update requests much more frequent (and
have those replace ping/pong) so that the communication failures become
apparent much sooner and the reconnection logic can kick in faster
(hopefully all in the short enough span of time that it wouldn't be
noticed by any interactive use.)

One notable problem right now is that every such reconnect loses the
list of commands that have been submitted. There is *some* feedback to
the user that this might be happening on the MQTT side, but nothing like
e.g. homie that makes the intent extremely obvious.